CALF/new_requirement.txt
absl-py==1.2.0
einops==0.4.1
opt-einsum==3.3.0
patool==1.15.0

PyWavelets==1.5.0
pytorch-wavelet==1.2

scikit-image==0.22.0
statsmodels==0.14.1
keopscore==2.1

transformers==4.38.2
peft==0.9.0
